+++ datamodel/openapi/openapi-generator/src/test/resources/DataModelGeneratorIntegrationTest/input-spec-with-builder/output/com/sap/cloud/sdk/services/builder/model/Soda.java
@@ -16,11 +16,11 @@
 
 package com.sap.cloud.sdk.services.builder.model;
 
+import java.util.Objects;
 import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.NoSuchElementException;
-import java.util.Objects;
 import java.util.Set;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -30,12 +30,10 @@ import com.fasterxml.jackson.annotation.JsonValue;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.io.Serializable;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.fasterxml.jackson.annotation.JsonTypeName;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -85,7 +83,8 @@ public class Soda
    * Get id
    * @return id  The id of this {@link Soda} instance.
    */
-  @Nonnull public Long getId() {
+  @Nonnull
+  public Long getId() {
     return id;
   }
 
@@ -113,7 +112,8 @@ public class Soda
    * Get name
    * @return name  The name of this {@link Soda} instance.
    */
-  @Nonnull public String getName() {
+  @Nonnull
+  public String getName() {
     return name;
   }
 
@@ -141,7 +141,8 @@ public class Soda
    * Get brand
    * @return brand  The brand of this {@link Soda} instance.
    */
-  @Nonnull public String getBrand() {
+  @Nonnull
+  public String getBrand() {
     return brand;
   }
 
@@ -181,7 +182,8 @@ public class Soda
    * Get flavors
    * @return flavors  The flavors of this {@link Soda} instance.
    */
-  @Nonnull public List<String> getFlavors() {
+  @Nonnull
+  public List<String> getFlavors() {
     return flavors;
   }
 
@@ -209,7 +211,8 @@ public class Soda
    * Get price
    * @return price  The price of this {@link Soda} instance.
    */
-  @Nonnull public Float getPrice() {
+  @Nonnull
+  public Float getPrice() {
     return price;
   }
 
@@ -305,7 +308,6 @@ public class Soda
     }
     return o.toString().replace("\n", "\n    ");
   }
-
     /**
     * Create a type-safe, fluent-api builder object to construct a new {@link Soda} instance with all required arguments.
     */