+++ datamodel/openapi/openapi-generator/src/test/resources/DataModelGeneratorIntegrationTest/input-spec-with-uppercase-file-extension/output/com/sap/cloud/sdk/services/uppercasefileextension/model/Soda.java
@@ -16,23 +16,21 @@
 
 package com.sap.cloud.sdk.services.uppercasefileextension.model;
 
+import java.util.Objects;
 import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.NoSuchElementException;
-import java.util.Objects;
 import java.util.Set;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.annotation.JsonValue;
-import java.io.Serializable;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.fasterxml.jackson.annotation.JsonTypeName;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -78,7 +76,8 @@ public class Soda
    * Get id
    * @return id  The id of this {@link Soda} instance.
    */
-  @Nonnull public Long getId() {
+  @Nonnull
+  public Long getId() {
     return id;
   }
 
@@ -106,7 +105,8 @@ public class Soda
    * Get name
    * @return name  The name of this {@link Soda} instance.
    */
-  @Nonnull public String getName() {
+  @Nonnull
+  public String getName() {
     return name;
   }
 
@@ -134,7 +134,8 @@ public class Soda
    * Get brand
    * @return brand  The brand of this {@link Soda} instance.
    */
-  @Nonnull public String getBrand() {
+  @Nonnull
+  public String getBrand() {
     return brand;
   }
 
@@ -162,7 +163,8 @@ public class Soda
    * Get flavor
    * @return flavor  The flavor of this {@link Soda} instance.
    */
-  @Nonnull public String getFlavor() {
+  @Nonnull
+  public String getFlavor() {
     return flavor;
   }
 
@@ -190,7 +192,8 @@ public class Soda
    * Get price
    * @return price  The price of this {@link Soda} instance.
    */
-  @Nonnull public Float getPrice() {
+  @Nonnull
+  public Float getPrice() {
     return price;
   }
 
@@ -286,7 +289,6 @@ public class Soda
     }
     return o.toString().replace("\n", "\n    ");
   }
-
 
 }