+++ datamodel/openapi/openapi-generator/src/test/resources/DataModelGeneratorIntegrationTest/input-spec-with-uppercase-file-extension/output/com/sap/cloud/sdk/services/uppercasefileextension/model/UpdateSoda.java
@@ -16,23 +16,21 @@
 
 package com.sap.cloud.sdk.services.uppercasefileextension.model;
 
+import java.util.Objects;
 import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.NoSuchElementException;
-import java.util.Objects;
 import java.util.Set;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.annotation.JsonValue;
-import java.io.Serializable;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.fasterxml.jackson.annotation.JsonTypeName;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -75,7 +73,8 @@ public class UpdateSoda
    * Get name
    * @return name  The name of this {@link UpdateSoda} instance.
    */
-  @Nonnull public String getName() {
+  @Nonnull
+  public String getName() {
     return name;
   }
 
@@ -103,7 +102,8 @@ public class UpdateSoda
    * Get brand
    * @return brand  The brand of this {@link UpdateSoda} instance.
    */
-  @Nonnull public String getBrand() {
+  @Nonnull
+  public String getBrand() {
     return brand;
   }
 
@@ -131,7 +131,8 @@ public class UpdateSoda
    * Get flavor
    * @return flavor  The flavor of this {@link UpdateSoda} instance.
    */
-  @Nonnull public String getFlavor() {
+  @Nonnull
+  public String getFlavor() {
     return flavor;
   }
 
@@ -159,7 +160,8 @@ public class UpdateSoda
    * Get price
    * @return price  The price of this {@link UpdateSoda} instance.
    */
-  @Nonnull public Float getPrice() {
+  @Nonnull
+  public Float getPrice() {
     return price;
   }
 
@@ -253,7 +255,6 @@ public class UpdateSoda
     }
     return o.toString().replace("\n", "\n    ");
   }
-
 
 }