+++ datamodel/openapi/openapi-generator/src/test/resources/DataModelGeneratorIntegrationTest/remove-operation-id-prefix/output/com/sap/cloud/sdk/services/builder/model/Order.java
@@ -16,23 +16,21 @@
 
 package com.sap.cloud.sdk.services.builder.model;
 
+import java.util.Objects;
 import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.NoSuchElementException;
-import java.util.Objects;
 import java.util.Set;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.annotation.JsonValue;
-import java.io.Serializable;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.fasterxml.jackson.annotation.JsonTypeName;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -78,7 +76,8 @@ public class Order
    * Get productId
    * @return productId  The productId of this {@link Order} instance.
    */
-  @Nonnull public Long getProductId() {
+  @Nonnull
+  public Long getProductId() {
     return productId;
   }
 
@@ -106,7 +105,8 @@ public class Order
    * Get quantity
    * @return quantity  The quantity of this {@link Order} instance.
    */
-  @Nonnull public Integer getQuantity() {
+  @Nonnull
+  public Integer getQuantity() {
     return quantity;
   }
 
@@ -134,7 +134,8 @@ public class Order
    * Get totalPrice
    * @return totalPrice  The totalPrice of this {@link Order} instance.
    */
-  @Nonnull public Float getTotalPrice() {
+  @Nonnull
+  public Float getTotalPrice() {
     return totalPrice;
   }
 
@@ -162,7 +163,8 @@ public class Order
    * Some typeless property, interpreted by the generator as nullable by default (because typeless)
    * @return typelessProperty  The typelessProperty of this {@link Order} instance.
    */
-  @Nullable public Object getTypelessProperty() {
+  @Nullable
+  public Object getTypelessProperty() {
     return typelessProperty;
   }
 
@@ -190,7 +192,8 @@ public class Order
    * Some typed property that is deliberately made nullable
    * @return nullableProperty  The nullableProperty of this {@link Order} instance.
    */
-  @Nullable public String getNullableProperty() {
+  @Nullable
+  public String getNullableProperty() {
     return nullableProperty;
   }
 
@@ -286,7 +289,6 @@ public class Order
     }
     return o.toString().replace("\n", "\n    ");
   }
-
 
 }