+++ datamodel/openapi/openapi-generator/src/test/resources/DataModelGeneratorIntegrationTest/remove-operation-id-prefix/output/com/sap/cloud/sdk/services/builder/model/SodaWithId.java
@@ -16,23 +16,21 @@
 
 package com.sap.cloud.sdk.services.builder.model;
 
+import java.util.Objects;
 import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.NoSuchElementException;
-import java.util.Objects;
 import java.util.Set;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.annotation.JsonValue;
-import java.io.Serializable;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.fasterxml.jackson.annotation.JsonTypeName;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -140,7 +138,8 @@ public class SodaWithId
    * Get name
    * @return name  The name of this {@link SodaWithId} instance.
    */
-  @Nonnull public String getName() {
+  @Nonnull
+  public String getName() {
     return name;
   }
 
@@ -168,7 +167,8 @@ public class SodaWithId
    * Get brand
    * @return brand  The brand of this {@link SodaWithId} instance.
    */
-  @Nonnull public String getBrand() {
+  @Nonnull
+  public String getBrand() {
     return brand;
   }
 
@@ -196,7 +196,8 @@ public class SodaWithId
    * Get quantity
    * @return quantity  The quantity of this {@link SodaWithId} instance.
    */
-  @Nonnull public Integer getQuantity() {
+  @Nonnull
+  public Integer getQuantity() {
     return quantity;
   }
 
@@ -224,7 +225,8 @@ public class SodaWithId
    * Get packaging
    * @return packaging  The packaging of this {@link SodaWithId} instance.
    */
-  @Nonnull public PackagingEnum getPackaging() {
+  @Nonnull
+  public PackagingEnum getPackaging() {
     return packaging;
   }
 
@@ -252,7 +254,8 @@ public class SodaWithId
    * Get price
    * @return price  The price of this {@link SodaWithId} instance.
    */
-  @Nonnull public Float getPrice() {
+  @Nonnull
+  public Float getPrice() {
     return price;
   }
 
@@ -280,7 +283,8 @@ public class SodaWithId
    * Get id
    * @return id  The id of this {@link SodaWithId} instance.
    */
-  @Nonnull public Long getId() {
+  @Nonnull
+  public Long getId() {
     return id;
   }
 
@@ -378,7 +382,6 @@ public class SodaWithId
     }
     return o.toString().replace("\n", "\n    ");
   }
-
 
 }